+++ src/Umber/Authentication/Resolver/CredentialResolverInterface.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Umber\Authentication\Resolver;
 
 use Umber\Authentication\AuthenticationMethodInterface;
-use Umber\Authentication\Exception\Resolver\CannotResolveAuthenticatedUserException;
+use Umber\Authentication\Exception\Resolver\CannotResolveAuthenticatedCredentialException;
 use Umber\Authentication\Exception\Resolver\CannotResolveAuthenticationMethodException;
 use Umber\Authentication\Exception\Resolver\UnsupportedAuthenticationMethodException;
 use Umber\Authentication\Exception\Token\TokenExpiredException;
@@ -24,7 +24,7 @@ interface CredentialResolverInterface
      * @throws TokenNotVerifiedException
      *
      * @throws CannotResolveAuthenticationMethodException
-     * @throws CannotResolveAuthenticatedUserException
+     * @throws CannotResolveAuthenticatedCredentialException
      * @throws UnsupportedAuthenticationMethodException
      */
     public function resolve(AuthenticationMethodInterface $method): CredentialInterface;